// ui/src/generated/rust-types/MailAccountOutput.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type MailAccountOutput = {
  id: string;
  displayName: string;
  email: string;
  provider: string;
  imapHost: string;
  imapPort: number;
  imapSecurity: string;
  smtpHost: string;
  smtpPort: number;
  smtpSecurity: string;
  senderName: string | null;
  isEnabled: boolean;
  syncInterval: number;
  lastSyncAt: string | null;
  createdAt: string;
};