// ui/src/generated/rust-types/MailFolderOutput.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type MailFolderOutput = {
  id: string;
  accountId: string;
  name: string;
  delimiter: string | null;
  folderType: string;
  totalCount: number;
  unreadCount: number;
  sortOrder: number;
};